+++ forge/src/main/java/com/jahirtrap/foodtxf/item/RecipeBookItem.java
@@ -1,9 +1,10 @@
 package com.jahirtrap.foodtxf.item;
 
+import com.klikli_dev.modonomicon.client.gui.BookGuiManager;
+import com.klikli_dev.modonomicon.client.gui.book.BookAddress;
 import net.minecraft.ChatFormatting;
 import net.minecraft.core.registries.BuiltInRegistries;
 import net.minecraft.network.chat.Component;
-import net.minecraft.server.level.ServerPlayer;
 import net.minecraft.world.InteractionHand;
 import net.minecraft.world.InteractionResult;
 import net.minecraft.world.entity.player.Player;
@@ -13,7 +14,6 @@ import net.minecraft.world.item.TooltipFlag;
 import net.minecraft.world.item.component.TooltipDisplay;
 import net.minecraft.world.level.Level;
 import net.minecraftforge.fml.ModList;
-import vazkii.patchouli.api.PatchouliAPI;
 
 import java.util.function.Consumer;
 
@@ -24,19 +24,16 @@ public class RecipeBookItem extends Item {
 
     @Override
     public InteractionResult use(Level level, Player player, InteractionHand hand) {
-        if (player instanceof ServerPlayer serverPlayer && ModList.get().isLoaded("patchouli"))
-            PatchouliAPI.get().openBookGUI(serverPlayer, BuiltInRegistries.ITEM.getKey(this));
-        else return super.use(level, player, hand);
-
-        return InteractionResult.SUCCESS;
+        if (ModList.get().isLoaded("modonomicon")) {
+            BookGuiManager.get().openBook(BookAddress.defaultFor(BuiltInRegistries.ITEM.getKey(this)));
+            return InteractionResult.SUCCESS;
+        } else return super.use(level, player, hand);
     }
 
     @Override
     public void appendHoverText(ItemStack stack, TooltipContext context, TooltipDisplay display, Consumer<Component> tooltip, TooltipFlag flag) {
-        if (!ModList.get().isLoaded("patchouli")) {
-            tooltip.accept(Component.translatable("tooltip.foodtxf.patchouli").withStyle(ChatFormatting.GRAY));
-        } else {
-            tooltip.accept(Component.translatable("patchouli.gui.lexicon.edition_str", "1st").withStyle(ChatFormatting.GRAY));
+        if (!ModList.get().isLoaded("modonomicon")) {
+            tooltip.accept(Component.translatable("tooltip.foodtxf.modonomicon").withStyle(ChatFormatting.GRAY));
         }
     }
 }